refactor: improving frontend folder structure
