+++ apps/frontend-web/src/app/app.routes.ts
@@ -1,18 +1,18 @@
 import { Route } from '@angular/router';
-import { authGuard } from './guards/auth.guard';
+import { authGuard } from './core/guards/auth.guard';
 
 export const appRoutes: Route[] = [
     { 
         path: 'login', 
-        loadComponent: () => import('./components/login/login').then(m => m.Login)
+        loadComponent: () => import('./features/auth/login/login').then(m => m.Login)
     },
     { 
         path: 'register', 
-        loadComponent: () => import('./components/register/register').then(m => m.Register)
+        loadComponent: () => import('./features/auth/register/register').then(m => m.Register)
     },
     { 
         path: 'dashboard', 
-        loadComponent: () => import('./components/dashboard/dashboard').then(m => m.Dashboard),
+        loadComponent: () => import('./features/files/dashboard/dashboard').then(m => m.Dashboard),
         canActivate: [authGuard]
     },
     { path: '', redirectTo: '/login', pathMatch: 'full' },
