+++ apps/frontend-web/src/app/core/guards/auth.guard.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { Auth } from '../services/auth';
+import { Auth } from '../auth/auth.service';
 
 export const authGuard: CanActivateFn = () => {
   const authService = inject(Auth);
