+++ apps/frontend-web/src/app/features/auth/register/register.spec.ts
@@ -1,23 +1,23 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { Login } from './login';
-import { Auth } from '../../services/auth';
+import { Register } from './register';
+import { Auth } from '../../../core/auth/auth.service';
 import { RouterTestingModule } from '@angular/router/testing';
 
-describe('Login', () => {
-  let component: Login;
-  let fixture: ComponentFixture<Login>;
+describe('Register', () => {
+  let component: Register;
+  let fixture: ComponentFixture<Register>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [Login, RouterTestingModule],
+      imports: [Register, RouterTestingModule],
       providers: [
-        { provide: Auth, useValue: { login: jest.fn() } as Partial<Auth> },
+        { provide: Auth, useValue: { register: jest.fn() } as Partial<Auth> },
       ],
     })
     .compileComponents();
 
-    fixture = TestBed.createComponent(Login);
+    fixture = TestBed.createComponent(Register);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
@@ -32,67 +32,66 @@ import { of, throwError } from 'rxjs';
 import { By } from '@angular/platform-browser';
 import { FormBuilder } from '@angular/forms';
 
-describe('Login - behaviors', () => {
-  let component: Login;
-  let fixture: ComponentFixture<Login>;
-  let loginMock: jest.Mock;
+describe('Register - behaviors', () => {
+  let component: Register;
+  let fixture: ComponentFixture<Register>;
+  let registerMock: jest.Mock;
 
   beforeEach(async () => {
-    loginMock = jest.fn().mockReturnValue(of({ accessToken: 'fake' }));
-    const authStub: Pick<Auth, 'login'> = {
-      login: loginMock as unknown as Auth['login'],
+    registerMock = jest.fn().mockReturnValue(of({ accessToken: 'fake' }));
+    const authStub: Pick<Auth, 'register'> = {
+      register: registerMock as unknown as Auth['register'],
     };
 
     await TestBed.configureTestingModule({
-      imports: [Login, RouterTestingModule],
+      imports: [Register, RouterTestingModule],
       providers: [{ provide: Auth, useValue: authStub }],
     }).compileComponents();
 
-    fixture = TestBed.createComponent(Login);
+    fixture = TestBed.createComponent(Register);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
 
-  it('should have submit button disabled when form invalid', () => {
+  it('should disable submit button when form invalid', () => {
     const btn: HTMLButtonElement = fixture.debugElement.query(By.css('button[type="submit"]')).nativeElement;
     expect(btn.disabled).toBe(true);
-    component.loginForm.patchValue({ email: 'user@example.com' });
+    component.registerForm.patchValue({ email: 'user@example.com' });
     fixture.detectChanges();
     expect(btn.disabled).toBe(true);
   });
 
-  it('should call Auth.login with credentials when form valid and submitted', () => {
-    component.loginForm.setValue({ email: 'user@example.com', password: 'secret' });
+  it('should call Auth.register when form valid and submitted', () => {
+    component.registerForm.setValue({ email: 'new@e.com', password: 'password123' });
     fixture.detectChanges();
 
     const form = fixture.debugElement.query(By.css('form'));
     form.triggerEventHandler('ngSubmit', {});
-    expect(loginMock).toHaveBeenCalledWith({ email: 'user@example.com', password: 'secret' });
+    expect(registerMock).toHaveBeenCalledWith({ email: 'new@e.com', password: 'password123' });
   });
 
-  it('should not call Auth.login when form is invalid', () => {
-    // form starts invalid by default
+  it('should not call Auth.register when form is invalid', () => {
     const form = fixture.debugElement.query(By.css('form'));
     form.triggerEventHandler('ngSubmit', {});
-    expect(loginMock).not.toHaveBeenCalled();
+    expect(registerMock).not.toHaveBeenCalled();
   });
 
-  it('should log error when Auth.login emits error', () => {
+  it('should log error when Auth.register emits error', () => {
     const consoleSpy = jest.spyOn(console, 'error').mockImplementation(() => undefined);
-    component.loginForm.setValue({ email: 'user@example.com', password: 'secret' });
-    loginMock.mockReturnValueOnce(throwError(() => new Error('bad')));
+    component.registerForm.setValue({ email: 'new@e.com', password: 'password123' });
+    registerMock.mockReturnValueOnce(throwError(() => new Error('bad')));
     const form = fixture.debugElement.query(By.css('form'));
     form.triggerEventHandler('ngSubmit', {});
     expect(consoleSpy).toHaveBeenCalled();
     consoleSpy.mockRestore();
   });
 
-  it('should not call Auth.login when form is valid but missing credentials', () => {
+  it('should not call Auth.register when form is valid but missing credentials', () => {
     // Replace the form with one without validators so it's valid even when empty
-    component.loginForm = new FormBuilder().group({ email: [''], password: [''] });
+    component.registerForm = new FormBuilder().group({ email: [''], password: [''] });
     fixture.detectChanges();
     const form = fixture.debugElement.query(By.css('form'));
     form.triggerEventHandler('ngSubmit', {});
-    expect(loginMock).not.toHaveBeenCalled();
+    expect(registerMock).not.toHaveBeenCalled();
   });
 });
