+++ apps/frontend-web/src/app/shared/dialogs/rename-dialog/rename-dialog.css
